Maktab50/requirements.txt
Django
django-jalali
psycopg2-binary
django-rest-framework
django-widget-tweaks